refactor: fix even more detekt issues
