+++ src/main/kotlin/org/cobalt/render/skia/gl/States.kt
@@ -16,7 +16,7 @@
  * You should have received a copy of the GNU General Public License along with Skija. If not, see <https://www.gnu.org/licenses/>.
  */
 
-package org.cobalt.util.skia.gl
+package org.cobalt.render.skia.gl
 
 import java.util.*
 import org.lwjgl.opengl.GL30.glGetIntegerv
